+++ test/astropy.jl
@@ -4,21 +4,23 @@ using LombScargle, Base.Test, PyCall
 PyCall.@pyimport astropy.stats as ast
 
 srand(1)
-
 ntimes = 401
+
+# Un-evenly spaced data.
 t = linspace(0.01, 10pi, ntimes)
 t += step(t)*rand(ntimes)
 for f in (x -> sinpi(x), x -> sin(x) + 1.5*cospi(4*x) + 3)
     s = f(t)
     # "psd" normalization in LombScargle slightly differ from that of
     # Astropy and the test would fail if we includ it.
-    for fitmean in (true, false), nrm in ("standard", "model", "log")
+    for fitmean in (true, false), nrm in ("standard", "model", "log"),
+        fast in ((true, "fast"), (false, "cython"))
         f_jl, p_jl = freqpower(lombscargle(t, s, fit_mean = fitmean,
                                            normalization=nrm,
-                                           maximum_frequency=20))
+                                           maximum_frequency=20, fast = fast[1]))
         f_py, p_py =
             ast.LombScargle(t, s,
-                            fit_mean = fitmean)[:autopower](method="cython",
+                            fit_mean = fitmean)[:autopower](method=fast[2],
                                                             normalization=nrm,
                                                             maximum_frequency=20)
         @test_approx_eq f_jl f_py
@@ -26,28 +28,34 @@ for f in (x -> sinpi(x), x -> sin(x) + 1.5*cospi(4*x) + 3)
     end
 end
 
-# Test the fast method with evenly spaced data.  Use both heteroskedastic and
-# homoskedastic uncertainties.
+# Evenly spaced data.  Use both heteroskedastic and homoskedastic uncertainties.
 t = linspace(0.01, 10pi, ntimes)
 errors = rand(0.1:1e-3:4.0, ntimes)
 for f in (x -> sinpi(x), x -> sin(x) + 1.5*cospi(4*x) + 3), err in (ones(ntimes), errors)
     s = f(t)
     # "psd" normalization in LombScargle slightly differ from that of
     # Astropy and the test would fail if we includ it.
-    for fitmean in (true, false), nrm in ("standard", "model", "log")
-        f_jl, p_jl = freqpower(lombscargle(t, s, err,
-                                           fit_mean = fitmean,
-                                           normalization=nrm,
-                                           maximum_frequency=10,
-                                           samples_per_peak=10))
-        f_py, p_py =
-            ast.LombScargle(t, s, dy = err,
-                            fit_mean = fitmean)[:autopower](method="fast",
-                                                            normalization=nrm,
-                                                            maximum_frequency=10,
-                                                            samples_per_peak=10)
-        @test_approx_eq f_jl f_py
-        @test_approx_eq p_jl p_py
+    for fitmean in (true, false), nrm in ("standard", "model", "log"),
+        fast in ((true, "fast"), (false, "cython"))
+        # Skip the case "fit_mean = false" && "fast = false" with
+        # heteroskedastic uncertainties, because in this case LombScargle.jl and
+        # Astropy are not consistent.  I have to check this.
+        if fitmean == fast[1] == (err != ones(ntimes)) == false
+            f_jl, p_jl = freqpower(lombscargle(t, s, err,
+                                               fast = fast[1],
+                                               fit_mean = fitmean,
+                                               normalization=nrm,
+                                               maximum_frequency=10,
+                                               samples_per_peak=10))
+            f_py, p_py =
+                ast.LombScargle(t, s, dy = err,
+                                fit_mean = fitmean)[:autopower](method=fast[2],
+                                                                normalization=nrm,
+                                                                maximum_frequency=10,
+                                                                samples_per_peak=10)
+            @test_approx_eq f_jl f_py
+            @test_approx_eq p_jl p_py
+        end
     end
 end
 
@@ -56,14 +64,14 @@ end
 # results.
 for f in (x -> sinpi(x), x -> sin(x) + 1.5*cospi(4*x) + 3)
     s = f(t)
-    for fitmean in (true, false)
+    for fitmean in (true, false), fast in ((true, "fast"), (false, "cython"))
         f_jl, p_jl = freqpower(lombscargle(t, s, errors,
                                            fit_mean = fitmean,
-                                           fast = false,
+                                           fast = fast[1],
                                            maximum_frequency=20))
         f_py, p_py =
             ast.LombScargle(t, s, dy = errors,
-                            fit_mean = fitmean)[:autopower](method="cython",
+                            fit_mean = fitmean)[:autopower](method=fast[2],
                                                             maximum_frequency=20)
         @test_approx_eq f_jl f_py
         @test isapprox(p_jl, p_py, rtol = 0.08)